+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Adding continuous integration with GitHub Actions (#28)
 - Adding unit tests for `annotate` command (#31)
 - Adding unit tests for `annotate-sv` command (#33)
+- Writing out gzip-ed files if output file name ends in `.gz` (#11)
 
 ## v0.21